backend/requirements.txt
flask
flask-cors
requests
python-dotenv
replicate
json